更改框架，采用pod方式，替换了大多数第三方框架
1.bug解决：other linker flag 加上$(inherited)；
2.MJRefresh、SVProgressHUD等出现问题，已解决
3.SVProgressHUD自定义还有一点小问题。
4.解决帐户页面点击进入二级页面不隐藏toolbar。

+++ SP2P_6.1/Network/Shove.OC/ConverUtil.m
@@ -20,7 +20,7 @@ static const char encodingTable[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopq
         char *characters = malloc(text.length*3/2);
     if (characters == NULL)
         return @"";
-        int end = text.length - 3;
+        int end = (int)text.length - 3;
     int index = 0;
     int charCount = 0;
     int n = 0;
